feat: Add functions to instantiate Gmail message and message part objects
This commit adds two new functions `instantiate_message_part` and `instantiate_message` to the `main.py` file.

The function `instantiate_message_part` takes a dictionary representing a Gmail message part and returns an object of type `GmailMessagePart` with the appropriate properties set. The function recursively creates child message parts if present.

The function `instantiate_message` takes a dictionary representing a Gmail message and returns an object of type `GmailMessage` with the appropriate properties set. It uses the `instantiate_message_part` function to create the payload object.

A new function `instantiate_gmail_message_list` is also added. This function takes a dictionary representing a response from the Gmail API and returns a list of Gmail message objects created using the `instantiate_message` function.

These functions are used in the `get_gmail_thread` function to create GmailThread objects with the messages property populated with instances of GmailMessage objects.

These changes improve the code by providing a more structured representation of Gmail messages and parts, making it easier to work with and manipulate the data.