feat: Add .gcloudignore file and Cloud Pub/Sub subscriber function
Added a new file `.gcloudignore` to the repository, which includes a reference to `.gitignore`. This file is used to specify patterns to exclude from Google Cloud deployments. Additionally, a new function `subscribe()` was added, which is triggered by a message on a Cloud Pub/Sub topic. This function prints out the data from Pub/Sub to verify its functionality.

The following changes were made:

- Added `.gcloudignore` file
- Updated `.gitignore` to ignore `.env*` and `.editorconfig`
- Modified `main.py` to include import statements for `cloudevents.http` and `functions_framework`
- Added `subscribe()` function for Cloud Pub/Sub event handling
- Updated `requirements.txt` to include `functions-framework` package